Refactoring: separate enum helpers
