+++ src/Events.h
@@ -6,166 +6,8 @@
 #include <Glacier/Enums.h>
 #include <Glacier/ZMath.h>
 #include "json.hpp"
+#include "Enums.h"
 #include "EventSystem.h"
-
-enum class MissionType {
-	Arcade,
-	Campaign,
-	Creation,
-	Elusive,
-	Escalation,
-	Evergreen,
-	Featured,
-	Mission,
-	Orbis,
-	Sniper,
-	Tutorial,
-	UserCreated,
-	VsRace,
-};
-
-enum class Events {
-	_47_FoundTrespassing,
-	AmbientChanged,
-	AccidentBodyFound,
-	Actorsick,
-	AddSyndicateTarget,
-	Agility_Start,
-	AllBodiesHidden,
-	BodyBagged,
-	BodyFound,
-	BodyHidden,
-	BrokenDisguiseCleared,
-	ContractStart,
-	CrowdNPC_Died,
-	DeadBodySeen,
-	Disguise,
-	DisguiseBlown,
-	Door_Unlocked,
-	Drain_Pipe_Climbed,
-	HoldingIllegalWeapon,
-	IntroCutEnd,
-	ItemDestroyed, // broken camcorder
-	ItemDropped,
-	ItemPickedUp,
-	ItemRemovedFromInventory,
-	ItemThrown,
-	Kill,
-	MurderedBodySeen,
-	Noticed_Pacified,
-	NoticedKill,
-	OpportunityEvents,
-	Pacify,
-	SecuritySystemRecorder,
-	setpieces,
-	ShotsFired,
-	SituationContained,
-	Spotted,
-	StartingSuit,
-	TargetBodySpotted,
-	TargetEliminated,
-	TargetEscapeFoiled, // Yuki killed in Gondola
-	Trespassing,
-	Unnoticed_Kill,
-	Unnoticed_Pacified,
-	Witnesses,
-};
-
-enum class SecuritySystemRecorderEvent {
-	Undefined,
-	Spotted,
-	Erased,
-	Destroyed,
-	CameraDestroyed,
-};
-
-inline auto getSecuritySystemRecorderEventFromString(const std::string& str) {
-	if (str == "spotted") return SecuritySystemRecorderEvent::Spotted;
-	if (str == "erased") return SecuritySystemRecorderEvent::Erased;
-	if (str == "destroyed") return SecuritySystemRecorderEvent::Destroyed;
-	if (str == "CameraDestroyed") return SecuritySystemRecorderEvent::CameraDestroyed;
-	return SecuritySystemRecorderEvent::Undefined;
-}
-
-inline auto getMissionTypeFromString(const std::string& str) -> std::optional<MissionType> {
-	static const std::map<std::string, MissionType> map = {{
-		{"arcade", MissionType::Arcade},
-		{"campaign", MissionType::Campaign},
-		{"elusive", MissionType::Elusive},
-		{"escalation", MissionType::Escalation},
-		{"evergreen", MissionType::Evergreen},
-		{"featured", MissionType::Featured},
-		{"mission", MissionType::Mission},
-		{"orbis", MissionType::Orbis},
-		{"sniper", MissionType::Sniper},
-		{"tutorial", MissionType::Tutorial},
-		{"usercreated", MissionType::UserCreated},
-		{"vsrace", MissionType::VsRace},
-	}};
-	auto const it = map.find(str);
-	if (it != map.end()) return it->second;
-	return std::nullopt;
-}
-
-inline auto getGameTensionFromValue(int n) -> EGameTension {
-	switch (static_cast<EGameTension>(n)) {
-	case EGameTension::EGT_Agitated: return EGameTension::EGT_Agitated;
-	case EGameTension::EGT_AlertedHigh: return EGameTension::EGT_AlertedHigh;
-	case EGameTension::EGT_AlertedLow: return EGameTension::EGT_AlertedLow;
-	case EGameTension::EGT_Ambient: return EGameTension::EGT_Ambient;
-	case EGameTension::EGT_Arrest: return EGameTension::EGT_Arrest;
-	case EGameTension::EGT_Combat: return EGameTension::EGT_Combat;
-	case EGameTension::EGT_Hunting: return EGameTension::EGT_Hunting;
-	case EGameTension::EGT_Searching: return EGameTension::EGT_Searching;
-	}
-	return EGameTension::EGT_Undefined;
-}
-
-inline auto getDeathContextFromValue(int n) -> EDeathContext {
-	switch (static_cast<EDeathContext>(n)) {
-	case EDeathContext::eDC_ACCIDENT: return EDeathContext::eDC_ACCIDENT;
-	case EDeathContext::eDC_HIDDEN: return EDeathContext::eDC_HIDDEN;
-	case EDeathContext::eDC_MURDER: return EDeathContext::eDC_MURDER;
-	case EDeathContext::eDC_NOT_HERO: return EDeathContext::eDC_NOT_HERO;
-	}
-	return EDeathContext::eDC_UNDEFINED;
-}
-
-inline auto getDeathTypeFromValue(int n) -> EDeathType {
-	switch (static_cast<EDeathType>(n)) {
-	case EDeathType::eDT_BLOODY_KILL: return EDeathType::eDT_BLOODY_KILL;
-	case EDeathType::eDT_KILL: return EDeathType::eDT_KILL;
-	case EDeathType::eDT_PACIFY: return EDeathType::eDT_PACIFY;
-	}
-	return EDeathType::eDT_UNDEFINED;
-}
-
-inline auto getKillTypeFromValue(int n) -> EKillType {
-	switch (static_cast<EKillType>(n)) {
-	case EKillType::EKillType_ChokeOut: return EKillType::EKillType_ChokeOut;
-	case EKillType::EKillType_Fiberwire: return EKillType::EKillType_Fiberwire;
-	case EKillType::EKillType_ItemTakeOutBack: return EKillType::EKillType_ItemTakeOutBack;
-	case EKillType::EKillType_ItemTakeOutFront: return EKillType::EKillType_ItemTakeOutFront;
-	case EKillType::EKillType_KnockOut: return EKillType::EKillType_KnockOut;
-	case EKillType::EKillType_PistolExecute: return EKillType::EKillType_PistolExecute;
-	case EKillType::EKillType_Pull: return EKillType::EKillType_Pull;
-	case EKillType::EKillType_Push: return EKillType::EKillType_Push;
-	case EKillType::EKillType_SnapNeck: return EKillType::EKillType_SnapNeck;
-	case EKillType::EKillType_Throw: return EKillType::EKillType_Throw;
-	}
-	return EKillType::EKillType_Undefined;
-}
-
-inline auto getActorTypeFromValue(int n) -> EActorType
-{
-	switch (static_cast<EActorType>(n))
-	{
-		case EActorType::eAT_Civilian: return EActorType::eAT_Civilian;
-		case EActorType::eAT_Guard: return EActorType::eAT_Guard;
-		case EActorType::eAT_Hitman: return EActorType::eAT_Hitman;
-	}
-	return EActorType::eAT_Last;
-}
 
 struct GameChanger {
 	//std::string Id;
